NIDAQ: Split DigitalOutputPort into its own file.
