+++ DeviceAdapters/NIDAQ/NIDAQ.vcxproj
@@ -92,6 +92,7 @@
   <ItemGroup>
     <ClCompile Include="NIDAQ.cpp" />
     <ClCompile Include="NIAnalogOutputPort.cpp" />
+    <ClCompile Include="NIDigitalOutputPort.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="NIDAQ.h" />
