+++ DeviceAdapters/NIDAQ/NIDAQ.vcxproj.filters
@@ -21,6 +21,9 @@
     <ClCompile Include="NIAnalogOutputPort.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="NIDigitalOutputPort.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="NIDAQ.h">
